+++ src/main/java/com/cropster/challenge/delval/repositories/StockRepository.java
@@ -1,5 +1,6 @@
 package com.cropster.challenge.delval.repositories;
 
+import java.math.BigDecimal;
 import java.util.Set;
 import javax.transaction.Transactional;
 import org.springframework.data.jpa.repository.Modifying;
@@ -25,6 +26,6 @@ public interface StockRepository extends CrudRepository<Stock, StockPK> {
   @Query(
       value = "UPDATE stock SET amount = amount - :amount WHERE facility_id = :facilityId and green_coffee_id = :greenCoffeeId and amount > 0",
       nativeQuery = true)
-  void updateAmount(Integer facilityId, Integer greenCoffeeId, Integer amount);
+  void updateAmount(Integer facilityId, Integer greenCoffeeId, BigDecimal amount);
 
 }
